create packages with source code and with test files. use pathlib to make sure code can access data files inside package properly
