+++ sdlc_xmalab/xma_data_processor.py
@@ -3,8 +3,11 @@
 import glob
 import os
 
+from subprocess import Popen
+
 import blend_modes
 import cv2
+from PIL import Image
 import numpy as np
 import pandas as pd
 from ruamel.yaml import YAML
@@ -104,6 +107,82 @@ class XMADataProcessor:
             raise FileNotFoundError(f'Empty trials directory found. Please put trials to be '
                                      'analyzed after training into the {path} folder')
         return trials
+
+    def split_rgb(self, trial_path, codec='avc1'):
+        '''Takes a RGB video with different grayscale data written to the R, G, and B channels and splits it back into its component source videos.'''
+        trial_name = os.path.basename(os.path.normpath(trial_path))
+        out_name = trial_name + '_split_'
+
+        try:
+            rgb_video = cv2.VideoCapture(f'{trial_path}/{trial_name}_rgb.avi')
+        except FileNotFoundError as e:
+            raise FileNotFoundError(f'Couldn\'t find video at {trial_path}/{trial_name}_rgb.avi') from e
+        frame_width = int(rgb_video.get(3))
+        frame_height = int(rgb_video.get(4))
+        frame_rate = round(rgb_video.get(5),2)
+        if codec == 'uncompressed':
+            pix_format = 'gray'   ##change to 'yuv420p' for color or 'gray' for grayscale. 'pal8' doesn't play on macs
+            cam1_split_ffmpeg = Popen(['ffmpeg', '-y', '-f', 'image2pipe', '-vcodec', 'png', '-r', str(int(frame_rate)),
+            '-i', '-', '-vcodec', 'rawvideo','-pix_fmt',pix_format,'-r', str(int(frame_rate)), f'{trial_path}/{out_name}'+'cam1.avi'], stdin=PIPE)
+            cam2_split_ffmpeg = Popen(['ffmpeg', '-y', '-f', 'image2pipe', '-vcodec', 'png', '-r', str(int(frame_rate)),
+            '-i', '-', '-vcodec', 'rawvideo','-pix_fmt',pix_format,'-r', str(int(frame_rate)), f'{trial_path}/{out_name}'+'cam2.avi'], stdin=PIPE)
+            blue_split_ffmpeg = Popen(['ffmpeg', '-y', '-f', 'image2pipe', '-vcodec', 'png', '-r', str(int(frame_rate)),
+            '-i', '-', '-vcodec', 'rawvideo','-pix_fmt',pix_format,'-r', str(int(frame_rate)), f'{trial_path}/{out_name}'+'blue.avi'], stdin=PIPE)
+        else:
+            if codec == 0:
+                fourcc = 0
+            else:
+                fourcc = cv2.VideoWriter_fourcc(*codec)
+            cam1 = cv2.VideoWriter(f'{trial_path}/{out_name}'+'cam1.avi',
+                                    fourcc,
+                                    frame_rate,(frame_width, frame_height))
+            cam2 = cv2.VideoWriter(f'{trial_path}/{out_name}'+'cam2.avi',
+                                    fourcc,
+                                    frame_rate,(frame_width, frame_height))
+            blue_channel = cv2.VideoWriter(f'{trial_path}/{out_name}'+'blue.avi',
+                                    fourcc,
+                                    frame_rate,(frame_width, frame_height))
+
+        i = 1
+        while rgb_video.isOpened():
+            ret, frame = rgb_video.read()
+            print(f'Current Frame: {i}')
+            i = i + 1
+            if ret:
+                B, G, R = cv2.split(frame)
+                if codec == 'uncompressed':
+                    im_r = Image.fromarray(R)
+                    im_g = Image.fromarray(G)
+                    im_b = Image.fromarray(B)
+                    im_r.save(cam1_split_ffmpeg.stdin, 'PNG')
+                    im_g.save(cam2_split_ffmpeg.stdin, 'PNG')
+                    im_b.save(blue_split_ffmpeg.stdin, 'PNG')
+                    if cv2.waitKey(1) & 0xFF == ord('q'):
+                        break
+                else:
+                    cam1.write(R)
+                    cam2.write(G)
+                    blue_channel.write(B)
+                    if cv2.waitKey(1) & 0xFF == ord('q'):
+                        break
+            else:
+                break
+        if codec == 'uncompressed':
+            cam1_split_ffmpeg.stdin.close()
+            cam1_split_ffmpeg.wait()
+            cam2_split_ffmpeg.stdin.close()
+            cam2_split_ffmpeg.wait()
+            blue_split_ffmpeg.stdin.close()
+            blue_split_ffmpeg.wait()
+        else:
+            cam1.release()
+            cam2.release()
+            blue_channel.release()
+        rgb_video.release()
+        cv2.destroyAllWindows()
+        print(f"Cam1 grayscale video created at {trial_path}/{out_name}cam1.avi!")
+        print(f"Cam2 grayscale video created at {trial_path}/{out_name}cam2.avi!")
+        print(f"Blue channel grayscale video created at {trial_path}/{out_name}blue.avi!")
 
     def _merge_rgb(self, trial_path, codec="avc1", mode="difference"):
         """
